Refactor frontend: split large components, lazy-load routes, extract styles (issue #20)
- Split BattlefieldPage (572 LOC) into useBattle hook + WarriorSelector,
  BattleControls, BattleStatus components (all under 100 LOC)
- Split BuilderPage (407 LOC) into useBuilder hook + WarriorListPanel,
  EditorToolbar, EditorStatus, WarriorListItem components (all under 100 LOC)
- Lazy-load BattlefieldPage, BuilderPage, and LearnPage via React.lazy +
  Suspense for code-split route chunks
- Extract all inline style={{ }} objects to module-level constants
- Add useMemo for derived library filter results (presets/userWarriors)
- Extract speed input onChange handlers to useCallback in BattleControls
