+++ frontend/src/main.tsx
@@ -1,12 +1,16 @@
-import React from 'react';
+import React, { Suspense } from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import './index.css';
 import AppLayout from './AppLayout';
 import HomePage from './pages/HomePage';
-import BattlefieldPage from './pages/BattlefieldPage';
-import BuilderPage from './pages/BuilderPage';
-import LearnPage from './pages/LearnPage';
+
+// eslint-disable-next-line react-refresh/only-export-components
+const BattlefieldPage = React.lazy(() => import('./pages/battlefield/BattlefieldPage'));
+// eslint-disable-next-line react-refresh/only-export-components
+const BuilderPage = React.lazy(() => import('./pages/builder/BuilderPage'));
+// eslint-disable-next-line react-refresh/only-export-components
+const LearnPage = React.lazy(() => import('./pages/LearnPage'));
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
@@ -14,9 +18,30 @@ ReactDOM.createRoot(document.getElementById('root')!).render(
       <Routes>
         <Route element={<AppLayout />}>
           <Route path="/" element={<HomePage />} />
-          <Route path="/battle" element={<BattlefieldPage />} />
-          <Route path="/builder" element={<BuilderPage />} />
-          <Route path="/learn" element={<LearnPage />} />
+          <Route
+            path="/battle"
+            element={
+              <Suspense>
+                <BattlefieldPage />
+              </Suspense>
+            }
+          />
+          <Route
+            path="/builder"
+            element={
+              <Suspense>
+                <BuilderPage />
+              </Suspense>
+            }
+          />
+          <Route
+            path="/learn"
+            element={
+              <Suspense>
+                <LearnPage />
+              </Suspense>
+            }
+          />
         </Route>
       </Routes>
     </BrowserRouter>
